refactor(starknet_client)!: fix imports and rename client structs
